+++ main.ts
@@ -1,4 +1,6 @@
 import "./paths";
 import { config } from "@packages/config";
 
-console.log(`🚀 (${config.env})`);
+console.log(
+  `🚀 http://${config.server.host}:${config.server.port} (${config.env})`,
+);